Updating JNI

+++ src/main/native/cpp/ctre_jni_mocks/BuffTrajPointStreamJNI.cpp
@@ -17,6 +17,7 @@ void* ConvertToWrapper(jlong aHandle)
 } // namespace
 
 extern "C" {
+
 /*
  * Class:     com_ctre_phoenix_motion_BuffTrajPointStreamJNI
  * Method:    Create1
@@ -50,7 +51,7 @@ JNIEXPORT jint JNICALL
 Java_com_ctre_phoenix_motion_BuffTrajPointStreamJNI_Destroy
   (JNIEnv*, jclass, jlong handle)
 {
-    return c_BuffTrajPointStream_Destroy(ConvertToWrapper(handle));
+    return (jint)c_BuffTrajPointStream_Destroy(ConvertToWrapper(handle));
 }
 
 /*
@@ -62,7 +63,7 @@ JNIEXPORT jint JNICALL
 Java_com_ctre_phoenix_motion_BuffTrajPointStreamJNI_Clear
   (JNIEnv*, jclass, jlong handle)
 {
-    return c_BuffTrajPointStream_Clear(ConvertToWrapper(handle));
+    return (jint)c_BuffTrajPointStream_Clear(ConvertToWrapper(handle));
 }
 
 /*
@@ -78,7 +79,7 @@ Java_com_ctre_phoenix_motion_BuffTrajPointStreamJNI_Write
    jint profileSlotSelect1, jboolean isLastPoint, jboolean zeroPos,
    jint timeDur, jboolean useAuxPID)
 {
-    return c_BuffTrajPointStream_Write(ConvertToWrapper(handle), position, velocity, arbFeedFwd, auxiliaryPos,
+    return (jint)c_BuffTrajPointStream_Write(ConvertToWrapper(handle), position, velocity, arbFeedFwd, auxiliaryPos,
             auxiliaryVel, auxiliaryArbFeedFwd, profileSlotSelect0, profileSlotSelect1, isLastPoint, zeroPos, timeDur, useAuxPID);
 }
